Move protocol types and class to /protocol/types
