+++ src/types/index.ts
@@ -1,11 +1,9 @@
-import Client from './client';
 import Server from './server';
 import Service from './service';
 import Workspace from './workspace';
 import WorkspacesController from '../controllers/workspaces';
 
 export {
-  Client,
   Service,
   Server,
   WorkspacesController,
