+++ src/utils/protocol/clientsList.ts
@@ -1,4 +1,4 @@
-import { Client } from '../types';
+import { Client } from './types';
 import ws from 'ws';
 
 /**
