+++ src/utils/protocol/types/index.ts
@@ -4,6 +4,7 @@ import { Message, MessagePayload } from './message';
 import OutgoingMessage from './outgoingMessage';
 import RequestMessage from './requestMessage';
 import ResponseMessage from './responseMessage';
+import Client from './client';
 
 export {
   TransportOptions,
@@ -12,5 +13,6 @@ export {
   Message,
   OutgoingMessage,
   ResponseMessage,
-  RequestMessage
+  RequestMessage,
+  Client
 };
